Refactored server side content plugin registry
