+++ server/content-plugin-registry/BaseRegistry.js
