+++ server/content-plugin-registry/elementRegistry.js
