+++ server/content-plugin-registry/index.js
@@ -1,0 +1,17 @@
+const containerRegistry = require('./containerRegistry');
+const elementRegistry = require('./elementRegistry');
+const Promise = require('bluebird');
+
+class ContentPluginRegistry {
+  constructor() {
+    this.containerRegistry = containerRegistry;
+    this.elementRegistry = elementRegistry;
+  }
+
+  initialize() {
+    const registires = Object.values(this);
+    return Promise.map(registires, it => it.initialize());
+  }
+}
+
+module.exports = new ContentPluginRegistry();
